move FileInput helpers to their own file
